>>Report model - id field added, - changed id to string which needs to be changed in DB >>CreateReport model created - fields with getters and setters >>ReportsController - changed to call repository, not gRPCclient - creates from "/reports" >>Report proto added CreateReportObject - rpc CreateReport(CreateReportObject) returns (ReportObject) - sets date,time,proof,description,status,location >>ReportsClient - Takes the - returns new Report with id and all the above >>File system - restructured repository
