+++ src/main/java/via/sep3/repository/intf/IReportsRepository.java
@@ -1,13 +1,11 @@
-package via.sep3.grpcclient.client;
+package via.sep3.repository.intf;
 
 import via.sep3.model.CreateReport;
 import via.sep3.model.Report;
 
 import java.util.List;
 
-public interface IReportsClient
-{
-
-    Report createReport(CreateReport newReport);
+public interface IReportsRepository {
     List<Report> getReports();
+    Report createReport(CreateReport report);
 }
